+++ src/pretty_printer.rs
@@ -230,6 +230,12 @@ impl<'a> PrettyPrinter<'a> {
         self
     }
 
+    /// Specify the maximum number of consecutive empty lines to print.
+    pub fn squeeze_empty_lines(&mut self, maximum: Option<usize>) -> &mut Self {
+        self.config.squeeze_lines = maximum;
+        self
+    }
+
     /// Specify the highlighting theme
     pub fn theme(&mut self, theme: impl AsRef<str>) -> &mut Self {
         self.config.theme = theme.as_ref().to_owned();